Move loading overlay examples to a Dev menu button.
